Move workouts-api files to its own directory
