+++ src/components/workouts-table/workouts-table-dialog/WorkoutsTableDialog.tsx
@@ -10,7 +10,7 @@ import type {
   WorkoutDetailsProps,
 } from '~/components/workouts-table/types'
 import { postWorkout, updateWorkout } from '~/api/workouts'
-import { invalidateGetWorkoutsQuery } from '~/api/workouts-helpers'
+import { invalidateGetWorkoutsQuery } from '~/api/workouts/workouts-helpers'
 import { useSnackbar } from '~/contexts/snackbar'
 import { useLoadingScreen } from '~/contexts/loading-screen'
